refactor: cleaned folder structure and added css
