+++ classes/Artwork.js
