+++ html/index.html
@@ -3,7 +3,11 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <script type="module" src="js/index.js" defer></script>
+    <script type="module" src="../js/index.js" defer></script>
+    <link rel="stylesheet" href="../css/master.css"/>
+    <link rel="stylesheet" href="../css/media.css"/>
+    <link rel="stylesheet" href="../css/reset.css"/>
+
     <title>Art Institute Chicago</title>
 </head>
 <body>
